requirements.txt
dill
einops
joblib
lightning
lion-pytorch
matplotlib
multiprocess
notebook
numpy
openpyxl
pandas
ptwt
PyWavelets
PyYAML
scikit-learn
scipy
seaborn
torch
torchinfo
torchmetrics
torchtuples
tqdm
wfdb